# implement a buffer optimized for writing to preallocated slices