--- a/internal/wire/blocked_frame_legacy.go
+++ b/internal/wire/blocked_frame_legacy.go
@@ -30,7 +30,7 @@ func parseBlockedFrameLegacy(r *bytes.Reader, _ protocol.VersionNumber) (Frame, 
 }
 
 //Write writes a BLOCKED frame
-func (f *blockedFrameLegacy) Write(b *bytes.Buffer, _ protocol.VersionNumber) error {
+func (f *blockedFrameLegacy) Write(b utils.ByteWriter, _ protocol.VersionNumber) error {
 	b.WriteByte(0x05)
 	utils.BigEndian.WriteUint32(b, uint32(f.StreamID))
 	return nil