--- a/internal/wire/frame.go
+++ b/internal/wire/frame.go
@@ -1,13 +1,12 @@
 package wire
 
 import (
-	"bytes"
-
 	"github.com/lucas-clemente/quic-go/internal/protocol"
+	"github.com/lucas-clemente/quic-go/internal/utils"
 )
 
 // A Frame in QUIC
 type Frame interface {
-	Write(b *bytes.Buffer, version protocol.VersionNumber) error
+	Write(b utils.ByteWriter, version protocol.VersionNumber) error
 	Length(version protocol.VersionNumber) protocol.ByteCount
 }