--- a/internal/wire/header.go
+++ b/internal/wire/header.go
@@ -87,7 +87,7 @@ func parsePacketHeader(b *bytes.Reader, sentBy protocol.Perspective, isPublicHea
 }
 
 // Write writes the Header.
-func (h *Header) Write(b *bytes.Buffer, pers protocol.Perspective, version protocol.VersionNumber) error {
+func (h *Header) Write(b utils.ByteWriter, pers protocol.Perspective, version protocol.VersionNumber) error {
 	if !version.UsesTLS() {
 		h.isPublicHeader = true // save that this is a Public Header, so we can log it correctly later
 		return h.writePublicHeader(b, pers, version)