--- a/internal/wire/stop_waiting_frame.go
+++ b/internal/wire/stop_waiting_frame.go
@@ -22,7 +22,7 @@ var (
 	errPacketNumberLenNotSet              = errors.New("StopWaitingFrame: PacketNumberLen not set")
 )
 
-func (f *StopWaitingFrame) Write(b *bytes.Buffer, v protocol.VersionNumber) error {
+func (f *StopWaitingFrame) Write(b utils.ByteWriter, v protocol.VersionNumber) error {
 	if v.UsesIETFFrameFormat() {
 		return errors.New("STOP_WAITING not defined in IETF QUIC")
 	}