@marten-seemann 2018-04-19
We'll find a better solution for the ACK frame, which doesn't require jumping around in the buffer.